Move SpacingSection into the ./styled folder
