+++ cardigan/stories/global/spacing/spacing.stories.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent } from 'react';
 import styled from 'styled-components';
 import Table from '@weco/common/views/components/Table/Table';
-import SpacingSection from '@weco/common/views/components/SpacingSection/SpacingSection';
+import SpacingSection from '@weco/common/views/components/styled/SpacingSection';
 import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
 import { PaletteColor, themeValues } from '@weco/common/views/themes/config';
