+++ common/views/components/styled/SpacingSection.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Space from '../styled/Space';
+import Space from './Space';
 
 const SpacingSection = styled(Space).attrs({
   v: { size: 'xl', properties: ['margin-bottom'] },
