+++ content/webapp/components/Body/CollectionsStaticContent.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent } from 'react';
 import Layout12 from '@weco/common/views/components/Layout12/Layout12';
 import SearchForm from '@weco/common/views/components/SearchForm/SearchForm';
-import SpacingSection from '@weco/common/views/components/SpacingSection/SpacingSection';
+import SpacingSection from '@weco/common/views/components/styled/SpacingSection';
 
 const CollectionsStaticContent: FunctionComponent = () => {
   return (
